// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/firmalar/models/firma-ekle.ts
export class FirmaEkle {
    Rol: string;
    BagliOlduguID: number;
    FirmaAd: string;
    FirmaKisaAd: string;
    Sifre: string;
    TelefonNumarasi: string;
    VergiTCNo: string;
    FaturaAdresi: string;
    Adres: string;
    Eposta: string;
    PsiBar: string;
    DisDerinligiSayisi: number;
    ParaBirimID: number;
    KayitTarihi: Date;
    KullaniciGorevi: string;
    KullaniciKisaKod: string;
    YetkiliKisi:string;
}